pyproject.toml
[tool.poetry]
name = "prometheus-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.0.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
channels = "^4.0.0"
daphne = "^4.1.0"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
